# Improve performance of ActiveSupport::JSON.encode